requirements.txt
langchain
pypdf
textract
python-dotenv==1.0.0
openai
# tiktoken